Create the board_presenter containing all the logic for putting the board together to be displayed
* board_presenter.go acts as a utility for console_ui.go to keep only the relevant parts of the UI in the lime light.
* The board_presenter contains the getDisplayBoard method and all methods that support it.
