+++ ui/board_presenter.go
@@ -1,0 +1,39 @@
+package ui
+
+import . "github.com/go-number-1-fan/tic-tac-toe/board"
+
+func getCharsPerRow(rowLength int) int {
+	charsPerRow := (rowLength * 3) + (rowLength + 1)
+	return charsPerRow
+}
+
+func getHorizontalDividerString(charsPerRow int) string {
+	boardString := " "
+	for j := 0; j < charsPerRow; j++ {
+		boardString = boardString + HorizontalDividerString
+	}
+	return boardString
+}
+
+func getBoardRowString(stringBoard []string, boardRowStartIndex int, rowLength int) string {
+	boardRow := ""
+	for boardColumn := 0; boardColumn < rowLength; boardColumn++ {
+		boardRow = boardRow + VerticalDividerString + stringBoard[boardRowStartIndex+boardColumn]
+	}
+	boardRow = boardRow + VerticalDividerString
+	return boardRow
+}
+
+func getDisplayBoard(board Board, player1Marker string, player2Marker string) string {
+	rowLength := board.RowLength()
+	charsPerRow := getCharsPerRow(rowLength)
+	boardString := NewLineString
+	stringBoard := board.StringBoard(player1Marker, player2Marker)
+	for rowStartIndex := 0; rowStartIndex < len(board); {
+		boardString = boardString + getHorizontalDividerString(charsPerRow) + NewLineString
+		boardString = boardString + getBoardRowString(stringBoard, rowStartIndex, rowLength) + NewLineString
+		rowStartIndex = rowStartIndex + rowLength
+	}
+	boardString = boardString + getHorizontalDividerString(charsPerRow) + NewLineString + NewLineString
+	return boardString
+}
